backend/requirements.txt
accelerate==1.12.0
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.11.0
async-timeout==4.0.3
attrs==25.4.0
av==16.0.1
brotlicffi @ file:///C:/miniconda3/conda-bld/brotlicffi_1762950448512/work
certifi @ file:///C:/miniconda3/conda-bld/certifi_1763135347537/work/certifi
cffi @ file:///C:/miniconda3/conda-bld/cffi_1761832794697/work
charset-normalizer @ file:///C:/miniconda3/conda-bld/charset-normalizer_1761744944645/work
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.4.13
contourpy==1.3.2
ctranslate2==4.6.1
cycler==0.12.1
distro==1.9.0
dnspython==2.8.0
elevenlabs==2.25.0
exceptiongroup==1.3.1
fastapi==0.123.9
faster-whisper==1.2.1
filelock @ file:///C:/miniconda3/conda-bld/filelock_1762420019633/work
flatbuffers==25.9.23
fonttools==4.60.1
frozenlist==1.8.0
fsspec==2025.10.0
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna @ file:///C:/miniconda3/conda-bld/idna_1761912044311/work
iniconfig==2.3.0
Jinja2 @ file:///C:/b/abs_920kup4e6u/croot/jinja2_1741711580669/work
jiter==0.12.0
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe @ file:///C:/b/abs_a0ma7ge0jc/croot/markupsafe_1738584052792/work
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
motor==3.7.1
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.7.0
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
numpy @ file:///C:/b/abs_0123vcxhf8/croot/numpy_and_numpy_base_1725470331966/work/dist/numpy-2.0.1-cp310-cp310-win_amd64.whl#sha256=e824ec3b279f4c5207736c83ac2f2dc79b489e08501a4ea9e3fc178c45db289b
ollama==0.6.1
onnxruntime==1.23.2
packaging==25.0
peft==0.18.0
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
pluggy==1.6.0
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
pycparser @ file:///C:/miniconda3/conda-bld/pycparser_1757496059965/work
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
pymongo==4.16.0
pyparsing==3.2.5
pypiwin32==223
pyreadline3==3.5.4
PySocks @ file:///C:/miniconda3/conda-bld/pysocks_1761753030270/work
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.20
pyttsx3==2.99
pytz==2025.2
pywin32==311
PyYAML @ file:///C:/miniconda3/conda-bld/pyyaml_1763465513320/work
regex==2025.11.3
requests @ file:///C:/miniconda3/conda-bld/requests_1762359612241/work
rich==14.2.0
safetensors==0.7.0
sentencepiece==0.2.1
shellingham==1.5.4
silero-vad==6.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
starlette==0.50.0
sympy==1.14.0
tokenizers==0.22.1
tomli==2.3.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
tqdm==4.67.1
transformers==4.57.1
typing-inspection==0.4.2
typing_extensions @ file:///C:/b/abs_ecq8gc0vbm/croot/typing_extensions_1756281142218/work
tzdata==2025.2
urllib3 @ file:///C:/b/abs_393g1ayee_/croot/urllib3_1750775481226/work
uvicorn==0.38.0
websockets==15.0.1
win_inet_pton @ file:///C:/miniconda3/conda-bld/win_inet_pton_1761746276406/work
yarl==1.22.0
